+++ extensions/positron-jupyter-debugger/src/runtimeNotebookDebugAdapterFactory.ts
@@ -10,6 +10,7 @@ import { log } from './extension.js';
 import { DebugCellManager } from './notebook.js';
 import { DisposableStore } from './util.js';
 import { JupyterRuntimeDebugAdapter } from './runtimeDebugAdapter.js';
+import { SourceMapper } from './sourceMapper.js';
 import { NotebookSourceMap } from './notebookSourceMap.js';
 
 const DEBUGGER_OUTPUT_CHANNEL_DESCRIPTOR = vscode.l10n.t('Debugger');
@@ -49,9 +50,9 @@ export class RuntimeNotebookDebugAdapterFactory implements vscode.DebugAdapterDe
 		// Create the output channel for the runtime session's debugger.
 		const outputChannel = this.createDebugAdapterOutputChannel(runtimeSession);
 
-		const adapter = this._disposables.add(new JupyterRuntimeDebugAdapter(outputChannel, debugSession, runtimeSession));
-		const sourceMap = this._disposables.add(new NotebookSourceMap(adapter, notebook));
-		adapter.setSourceMap(sourceMap);
+		const sourceMapper = new SourceMapper();
+		const sourceMap = this._disposables.add(new NotebookSourceMap(sourceMapper, notebook));
+		const adapter = this._disposables.add(new JupyterRuntimeDebugAdapter(sourceMapper, sourceMap, outputChannel, debugSession, runtimeSession));
 
 		// Create a debug cell manager to handle the cell execution and debugging.
 		const debugCellManager = this._disposables.add(new DebugCellManager(adapter, debugSession, notebook, runtimeSession, cell.index));
